move util tests
