+++ test/js/components/utils/labelWithSpinner.test.jsx
@@ -1,0 +1,12 @@
+import { render } from '@testing-library/react';
+import React from 'react';
+
+import { LabelWithSpinner } from '@/components/utils';
+
+describe('<LabelWithSpinner />', () => {
+  test('renders with defaults', () => {
+    const { getByText } = render(<LabelWithSpinner label="testing" />);
+
+    expect(getByText('testing')).toBeVisible();
+  });
+});
